Refactor ImprovedSessionStats component by modularizing sections into individual components for better organization and maintainability. Introduce new sections for session configuration, participants, targets, engagements, and summary. Update type definitions and enhance the overall structure of the session statistics form. Add comprehensive documentation for the new component structure.
